cms部分全部开发完成

+++ jeegem-service-admin/src/main/java/com/jeegem/service/ArticleService.java
@@ -69,6 +69,6 @@ public interface ArticleService {
 	 * @param id
 	 * @return
 	 */
-	Article getArticleById(Long id);
+	Article getArticleById(Integer id);
 
 }
